% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_user_world}
\alias{check_user_world}
\title{Check user's own world}
\usage{
check_user_world(world, lang)
}
\arguments{
\item{world}{The world provided by the user. It's a list. See details in
\code{\link{generar_mundo}}.}
}
\value{
If a misespecification is found, this function produces a stop and
  provides a descriptive error message.
}
\description{
This function analyzes if a world provided by the user satisfies all
requirements.
}
\details{
This function is called by \code{\link{.generate_world}}.
}
\keyword{internal}
