% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_walls}
\alias{check_walls}
\title{Check the walls user's provided world}
\usage{
check_walls(dataset, name, nx, ny, lang)
}
\arguments{
\item{dataset}{Either hor_walls or ver_walls}

\item{name}{Character string: "hor_walls" or "ver_walls"}

\item{nx, }{ny Size of the world}

\item{lang}{language code (such as "en" or "es") for printing messages}
}
\value{
If a misespecification is found, this function produces a stop and
  provides a descriptive error message.
}
\description{
This is a helper function to check ver_walls and hor_walls. It's called by
check_user_world twice.
}
\keyword{internal}
