% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{condiciones}
\alias{condiciones}
\alias{frente_abierto}
\alias{frente_cerrado}
\alias{izquierda_abierto}
\alias{izquierda_cerrado}
\alias{derecha_abierto}
\alias{derecha_cerrado}
\alias{hay_cosos}
\alias{no_hay_cosos}
\alias{karel_tiene_cosos}
\alias{karel_no_tiene_cosos}
\alias{mira_al_este}
\alias{mira_al_norte}
\alias{mira_al_oeste}
\alias{mira_al_sur}
\title{Condiciones que Karel puede verificar}
\usage{
frente_abierto()

frente_cerrado()

izquierda_abierto()

izquierda_cerrado()

derecha_abierto()

derecha_cerrado()

hay_cosos()

no_hay_cosos()

karel_tiene_cosos()

karel_no_tiene_cosos()

mira_al_este()

mira_al_norte()

mira_al_oeste()

mira_al_sur()
}
\value{
Valor lógico TRUE o FALSE.
}
\description{
Este conjunto de funciones devuelven un valor lógico \code{TRUE} o
\code{FALSE} según la evaluación que Karel puede hacer de su mundo.
}
\details{
Las funciones \code{frente_abierto()}, \code{frente_cerrado()},
  \code{izquierda_abierto()}, \code{izquierda_cerrado()},
  \code{derecha_abierto()} y \code{derecha_cerrado()} analizan si hay paredes
  al frente, a la izquierda o a la derecha de Karel. Las funciones
  \code{hay_cosos()} y \code{no_hay_cosos()} analizan si hay \code{cosos} en
  la posición actual de Karel. Las funciones \code{karel_tiene_cosos()} y
  \code{karel_no_tiene_cosos()} analizan si Karel tiene \code{cosos} en su
  mochila (no visibles en la representación gráfica). Las funciones
  \code{mira_al_este()}, \code{mira_al_oeste()}, \code{mira_al_norte()} y
  \code{mira_al_sur()} analizan la dirección hacia la cual Karel está
  mirando.
}
\examples{
generar_mundo("mundo001")
frente_abierto()
frente_cerrado()
izquierda_abierto()
izquierda_cerrado()
derecha_abierto()
derecha_cerrado()
hay_cosos()
no_hay_cosos()
karel_tiene_cosos()
karel_no_tiene_cosos()
mira_al_este()
mira_al_oeste()
mira_al_norte()
mira_al_sur()

}
\seealso{
\code{\link{generar_mundo}}
}
