% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_en.R, R/lang_wrappers_es.R
\name{conditions}
\alias{conditions}
\alias{front_is_clear}
\alias{front_is_blocked}
\alias{left_is_clear}
\alias{left_is_blocked}
\alias{right_is_clear}
\alias{right_is_blocked}
\alias{beepers_present}
\alias{no_beepers_present}
\alias{karel_has_beepers}
\alias{karel_has_no_beepers}
\alias{facing_east}
\alias{facing_west}
\alias{facing_north}
\alias{facing_south}
\alias{frente_abierto}
\alias{frente_cerrado}
\alias{izquierda_abierto}
\alias{izquierda_cerrado}
\alias{derecha_abierto}
\alias{derecha_cerrado}
\alias{hay_cosos}
\alias{no_hay_cosos}
\alias{karel_tiene_cosos}
\alias{karel_no_tiene_cosos}
\alias{mira_al_este}
\alias{mira_al_oeste}
\alias{mira_al_norte}
\alias{mira_al_sur}
\title{Conditions that Karel can test}
\usage{
front_is_clear()

front_is_blocked()

left_is_clear()

left_is_blocked()

right_is_clear()

right_is_blocked()

beepers_present()

no_beepers_present()

karel_has_beepers()

karel_has_no_beepers()

facing_east()

facing_west()

facing_north()

facing_south()

frente_abierto()

frente_cerrado()

izquierda_abierto()

izquierda_cerrado()

derecha_abierto()

derecha_cerrado()

hay_cosos()

no_hay_cosos()

karel_tiene_cosos()

karel_no_tiene_cosos()

mira_al_este()

mira_al_oeste()

mira_al_norte()

mira_al_sur()
}
\value{
Logical value TRUE or FALSE.
}
\description{
These group of functions return a logical value \code{TRUE} o \code{FALSE} according to Karel's evaluation of her world.
}
\details{
The functions \code{front_is_clear()}, \code{front_is_blocked()}, \code{left_is_clear()}, \code{left_is_blocked()}, \code{right_is_clear()} y \code{right_is_blocked()} test if there is a wall in front of Karel, to her left or to her right, respectively. The functions \code{beepers_present()} y \code{no_beepers_present()} test if there are or there are not \code{beepers} at Karel's current position. The functions \code{karel_has_beepers()} y \code{karel_has_no_beepers()} test if Karel has or hasn't got  \code{beepers} in her bag (not visible in the plot). The functions \code{facing_east()}, \code{facing_west()}, \code{facing_north()} y \code{facing_south()} test the direction to which Karel is facing right now.
}
\examples{
generate_world("mundo001")
front_is_clear()
front_is_blocked()
left_is_clear()
left_is_blocked()
right_is_clear()
right_is_blocked()
beepers_present()
no_beepers_present()
karel_has_beepers()
karel_has_no_beepers()
facing_east()
facing_west()
facing_north()
facing_south()

}
\seealso{
\code{\link{generate_world}}
}
