% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{create_beepers}
\alias{create_beepers}
\title{Create dataset about beepers}
\usage{
create_beepers(nx = NULL, pos_x = NULL, pos_y = NULL, n = NULL, moment = 1)
}
\arguments{
\item{nx}{horizontal size of the world}

\item{pos_x, }{pos_y vectors of coordinates of cells with non-zero amount of
beepers}

\item{n}{number of beepers in each cell indicated by coordinates in
\code{pos_x} and \code{pos_y}}

\item{moment}{time}
}
\value{
A tibble with as many rows as cells with beepers in the world and 5
  columns: \code{x} and \code{y} for the coordinates of the cell, \code{cell}
  is the number of the cell counting as cell number 1 the cell in the bottom
  left corner and going upwards by row (meaning cell number 2 would be the
  cell in coordinates x=2 and y=1), \code{n} the number of beepers in this
  cell and \code{moment} the moment in which this state of the world
  corresponds to.
}
\description{
Given the elements provided in the world, this function generates a dataset
with info about the beepers present in the world. This function is called
from \code{generar_mundo()}.
}
\keyword{internal}
