% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_pkg_env}
\alias{.get_pkg_env}
\title{Get Karel's environment}
\usage{
.get_pkg_env()
}
\value{
An enviroment with objects that represent Karel's world.
}
\description{
This function returns the environment called pkg_env created by the package.
It's useful for debugging and checking. It's an internal function.
}
\details{
\code{pkg_env} is an environment created inside the package to store
  and share between functions all the objects related to Karel's world and
  its state. Since the functions that will be used by the students should be
  simple and without arguments (for example, \code{move()}), these functions
  modify internally \code{pkg_env}.

  The components of this environment are:
  \enumerate{
    \item \code{nx}: TODO
    \item \code{ny}:
    \item \code{hor_walls}:
    \item \code{ver_walls}:
    \item \code{open_moves}:
    \item \code{karel}:
    \item \code{dir_now}:
    \item \code{x_now}:
    \item \code{y_now}:
    \item \code{moment}:
    \item \code{beepers_any}:
    \item \code{beepers_bag}:
    \item \code{beepers_now}:
    \item \code{beepers_all}:
    \item \code{base_plot}:
  }
}
\keyword{internal}
