% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_karel.R
\name{.load_super_karel}
\alias{.load_super_karel}
\title{Turn on Karel's superpowers}
\usage{
.load_super_karel()
}
\description{
After running \code{.load_super_karel()}, Karel can also turn right and turn
around with \code{.turn_right()} and \code{.turn_around()}. If these
superpowers aren't loaded, then these functions won't be available and Karel
can't use them. This is an internal function, and its called by language
wrappers.
}
\keyword{internal}
