% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{.run_actions}
\alias{.run_actions}
\title{Run actions}
\usage{
.run_actions(loop = TRUE, lang)
}
\arguments{
\item{loop}{A logical value TRUE or FALSE indicating if the animation should
repeat itself after finished or not (defaults to TRUE).}
}
\value{
Produces an animation with \code{gganimate}.
}
\description{
This function produces the animation that shows all actions performed by
Karel since its world was generated by \code{generate_world}. This is
internal, it's called by a wrapper depending on user's language
}
\keyword{internal}
