% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_beepers_df_row}
\alias{get_beepers_df_row}
\title{Get row number in beepers dataset}
\usage{
get_beepers_df_row()
}
\value{
row number or numeric of length 0
}
\description{
Given current position of Karel, it calculates the cell id (\code{cell}) and
returns the row number of the respective cell in the beepers_now dataset.
It's used to add or remove beepers in this row when Karel picks or puts
beepers.
}
\keyword{internal}
