% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_karel.R
\name{global_env_set_hack}
\alias{global_env_set_hack}
\title{This function lets me add objects to another environment}
\usage{
global_env_set_hack(key, val, pos)
}
\description{
This function lets me add objects to another environment
}
