% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_es.R
\name{graficar_mundo_estatico}
\alias{graficar_mundo_estatico}
\title{Producir un gráfico del mundo de Karel en un momento dado}
\usage{
graficar_mundo_estatico(momento)
}
\arguments{
\item{momento}{El momento que se desea graficar.}
}
\value{
Imprime el gráfico.
}
\description{
Esta función grafica el mundo de Karel en el momento pedido. Inicialmente,
momento toma el valor 1 y con cada acción que Karel realiza se incrementa en
1. El momento actual está guardado en \code{pkg_env$moment}. Esta función es
útil para revisar el código y para obtener imágenes estáticas que pueden
usarse al crear ejemplos y ejercicios en los materiales de estudio para los
estudiantes. Es una función interna, no está pensada para ser usada por
estudiantes, pero se puede usar con karel::graficar_mundo_estatico().
}
