% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_en.R
\name{load_super_karel}
\alias{load_super_karel}
\title{Turn on Karel's superpowers}
\usage{
load_super_karel()
}
\value{
It doesn't return anything but attaches to the global environment the
  functions \code{turn_right()} and \code{turn_around()}.
}
\description{
After running \code{load_super_karel()}, Karel can also turn right and turn
around with \code{turn_right()} and \code{turn_around()}. If these
superpowers aren't loaded, then these functions won't be available and Karel
can't use them.
}
\examples{
generate_world("mundo001")
load_super_karel()
turn_around()
turn_right()
if (FALSE) run_actions()

}
\seealso{
\code{\link{actions}} \code{\link{generate_world}}
  \code{\link{run_actions}}
}
