% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_base_world}
\alias{plot_base_world}
\title{Plot base world}
\usage{
plot_base_world()
}
\value{
A ggplot with the world
}
\description{
This functions produces the initial plot of the world, with its size and all
the walls if there are any. It doesn't plot Karel or the beepers, since those
things can change with time.
}
\keyword{internal}
