% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_en.R
\name{plot_static_world}
\alias{plot_static_world}
\title{Plot the world at a given time}
\usage{
plot_static_world(time)
}
\arguments{
\item{time}{The requested time}
}
\value{
Prints the plot.
}
\description{
This function plots Karel'w wort at the requested time. Initially, time is 1
and with each action that Karel performs, time is incremented by one. Current
time is stored in \code{pkg_env$moment}. This function is useful for
debugging and to get static images to be used in the examples in the handouts
for students. It's an internal function, not thought to be used by students,
but can be used with karel:::plot_static_world().
}
\examples{
if (interactive()) karel:::plot_static_world(1)

}
