% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{put_hor_walls}
\alias{put_hor_walls}
\title{Put horizontal walls (streets)}
\usage{
put_hor_walls(x, y, lgth)
}
\arguments{
\item{x, }{y x and y coordinates of the beginning point of the wall}

\item{lgth}{length of the wall}
}
\value{
A lgth x 3 dataset with coordinates where Karel can't move and the
  direction (side) towards it can't move because of this piece of horizontal
  wall.
}
\description{
Helper function for \code{generate_open_moves()}. It takes the available data
for a piece of horizontal wall (x and y coordinates of its beginning point
and its length) and produces a dataset with the positions and directions of
prohibited moves for Karel because of this wall.
}
\keyword{internal}
