% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_karel.R
\name{subirfcn2}
\alias{subirfcn2}
\title{Title}
\usage{
subirfcn2()
}
\value{

}
\description{
Con esta queda en el globa env la funcion llamada .funcioninterna,
pero como tiene el punto inicial no se, anda, corriendo .funcioninterna()
}
