% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_karel.R
\name{subirfcn3}
\alias{subirfcn3}
\title{Title}
\usage{
subirfcn3()
}
\value{

}
\description{
Con esta queda en el globa env la funcion llamada .funcioninterna2,
pero como tiene el punto inicial no se, anda, corriendo .funcioninterna2()
}
