% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_en.R
\name{actions}
\alias{actions}
\alias{move}
\alias{turn_left}
\alias{put_beeper}
\alias{pick_beeper}
\alias{turn_right}
\alias{turn_around}
\title{Available actions for Karel}
\usage{
move()

turn_left()

put_beeper()

pick_beeper()

turn_right()

turn_around()
}
\value{
These functions don't return anything, but make changes in Karel's
  world that are visible when all the actions are run through
  \code{run_actions()}.
}
\description{
\code{move()}, \code{turn_left()}, \code{pick_beeper()} y \code{put_beeper()}
are the four basic activities that Karel can perform. If you turn on Karel's
superpowers with \code{load_super_karel()}, then she can also
\code{turn_right()} y \code{turn_around()}.
}
\examples{
generate_world("mundo001")
move()
pick_beeper()
turn_left()
put_beeper()
if (interactive()) run_actions()

}
\seealso{
\code{\link{load_super_karel}} \code{\link{generate_world}}
  \code{\link{run_actions}}
}
