% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_es.R
\name{conseguir_amb}
\alias{conseguir_amb}
\title{Obtener el ambiente de Karel}
\usage{
conseguir_amb()
}
\value{
Un ambiente de R con objetos que representan al mundo de Karel.
}
\description{
Esta funci칩n devuelve un ambiente (R environment) llamado pkg_env, que es
creado por el paquete. Se puede usar para probar el funcionamiento del
paquete. Es una funci칩n interna, no est치 pensada para ser usada por
estudiantes, pero se puede usar con karel::conseguir_amb().
}
\details{
\code{pkg_env} es un ambiente de R creado dentro del paquete para
  guardar y compartir entre las funciones todos los objetos relacionados con
  el mundo de Karel y su estado en cada momento. Dado que estas funciones que
  usan los estudiantes deben ser simples y no usar argumentos (como, por
  ejemplo, \code{avanzar()}) estas funciones modifican internamente a
  \code{pkg_env} para implementar cada acci칩n.

  Los componentes de este ambiente son:
  \enumerate{
    \item \code{nx}: TODO
    \item \code{ny}:
    \item \code{hor_walls}:
    \item \code{ver_walls}:
    \item \code{open_moves}:
    \item \code{karel}:
    \item \code{dir_now}:
    \item \code{x_now}:
    \item \code{y_now}:
    \item \code{moment}:
    \item \code{beepers_any}:
    \item \code{beepers_bag}:
    \item \code{beepers_now}:
    \item \code{beepers_all}:
    \item \code{base_plot}:
  }
}
