% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{generar_mundo}
\alias{generar_mundo}
\title{Generar el mundo de Karel}
\usage{
generar_mundo(mundo)
}
\arguments{
\item{mundo}{Un caracter de largo 1 indicando el nombre de uno de los mundos
que ya vienen en el paquete o un objeto de tipo lista con todos los
componentes que debe tener un mundo (ver más abajo en Detalles).}
}
\value{
Dibuja el estado inicial del mundo de Karel y deja todo preparado
  para comenzar a registrar sus acciones.
}
\description{
Esta función toma la definición de un mundo para Karel (tamaño, paredes,
"cosos" e incluso la ubicación y dirección de Karel) y lo muestra en un
gráfico. Además, prepara todo para que Karel pueda realizar sus acciones, de
modo que debe ser evaluada antes de que Karel empiece a cumplir sus
objetivos. En especial, si en algún momento hemos cometido un error, debemos
comenzar de nuevo corriendo esta función.
}
\details{
Luego de correr \code{generar_mundo()}, se ejecutan las acciones de Karel y
se pueden visualizar con la función \code{ejecutar_acciones()}.

El argumento \code{mundo} puede consistir de un mundo creado (es
  decir, inventado) por cualquiera. En este caso, \code{mundo} debe ser una
  lista con los siguientes componentes:

  \enumerate{
    \item \code{nx}: TODO
    \item \code{ny}:
    \item \code{hor_walls}:
    \item \code{ver_walls}:
    \item \code{karel_x}:
    \item \code{karel_y}:
    \item \code{karel_dir}:
    \item \code{beepers_x}:
    \item \code{beepers_y}:
    \item \code{beepers_n}:
    \item \code{beepers_bag}:
  }
}
\examples{
generar_mundo("world_101")

}
\seealso{
\code{\link{acciones}} \code{\link{ejecutar_acciones}}
}
