% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_karel.R
\name{global_env_set_hack}
\alias{global_env_set_hack}
\title{This function lets me add objects to another environment}
\usage{
global_env_set_hack(key, val, pos)
}
\arguments{
\item{key}{name for the new object}

\item{val}{the object}

\item{pos}{the index for the environment}
}
\description{
This function lets me add objects to another environment
}
\keyword{internal}
