% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_static_world}
\alias{plot_static_world}
\title{Plot the world at a given time}
\usage{
plot_static_world(time)
}
\arguments{
\item{time}{The requested time}
}
\value{
Prints the plot.
}
\description{
This function plots Karel'w wort at the requested time. Initially, time is 1
and with each action that Karel performs, time is incremented by one. Current
time is stored in \code{pkg_env$moment}. This function is useful for
debuggint and to get static images to be used in the examples in the handouts
for studentes.
}
