% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_wrappers_en.R
\name{run_actions}
\alias{run_actions}
\title{Run actions}
\usage{
run_actions(loop = FALSE)
}
\arguments{
\item{loop}{A logical value TRUE or FALSE indicating if the animation should
repeat itself after finished or not (defaults to FALSE).}
}
\value{
Produces an animation with \code{gganimate}.
}
\description{
This function produces the animation that shows all actions performed by
Karel since its world was generated by \code{generate_world}.
}
\examples{
generate_world("mundo001")
move()
pick_beeper()
turn_left()
put_beeper()
if (FALSE) run_actions()

}
\seealso{
\code{\link{generate_world}}
}
